moved client out of index export
