+++ server/db/client.js
@@ -2,7 +2,6 @@ require("dotenv").config("../../.env");
 const { Client } = require("pg");
 const client = new Client(process.env.DATABASE_URL);
 
-console.log(process.env.DATABASE_URL);
 
 module.exports = {
     client,
