+++ server/db/orders.js
@@ -1,4 +1,4 @@
-const { client } = require("./index");
+const { client } = require("./client");
 
 // orders database functions
